Modify EnabledByAgreementVersion from decimal to int

+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Models/FeatureToggles/Feature.cs
@@ -1,6 +1,6 @@
 ﻿namespace SFA.DAS.EAS.Domain.Models.FeatureToggles
 {
-    public class Feature
+	public class Feature
     {
         public bool Enabled { get; set; }
 
@@ -8,7 +8,7 @@
         ///     The version of the agreement that introduced this feature. Signing this or any subsequent
         ///     agreement entitles access to this feature.
         /// </summary>
-        public decimal EnabledByAgreementVersion { get; set; }
+        public int EnabledByAgreementVersion { get; set; }
 
         public FeatureType FeatureType { get; set; }
         public string Name { get; set; }
